Change: 設定ファイルの読み込み: cli.go -> app.go
