+++ app.go
@@ -1,9 +1,13 @@
 package main
 
 import (
+	"bufio"
 	"fmt"
+	"log"
 	"os"
 	"time"
+
+	homedir "github.com/mitchellh/go-homedir"
 )
 
 const (
@@ -76,6 +80,47 @@ func (dt *Dt) String() string {
 	default:
 		return t.Format(f)
 	}
+}
+
+func loadConfig() {
+	path, err := homedir.Expand("~/.config/dt/.dt")
+	if err != nil {
+		return
+	}
+
+	f, err := os.Open(path)
+	if err != nil {
+		return
+	}
+	defer f.Close()
+
+	scanner := bufio.NewScanner(f)
+	m := map[string]string{}
+	for scanner.Scan() {
+		k, v := SplitFormat(scanner.Text())
+		if k == "" || v == "" {
+			continue
+		}
+		m[k] = v
+	}
+
+	for k, v := range m {
+		log.Printf("custom format: %s => %s\n", k, v)
+		formats[k] = v
+	}
+}
+
+func SplitFormat(s string) (string, string) {
+	cols := splitRegexp.Split(s, 2)
+	if len(cols) != 2 {
+		return "", ""
+	}
+
+	if cols[0] == "" || cols[1] == "" {
+		return "", ""
+	}
+
+	return cols[0], cols[1]
 }
 
 func main() {
