+++ app_test.go
@@ -80,3 +80,22 @@ func TestDt_AddSecond(t *testing.T) {
 		t.Errorf("Dt.AddSecond() = %v, want %v", actual, expect)
 	}
 }
+
+func TestSplitFormat(t *testing.T) {
+	params := []struct {
+		input string
+		key   string
+		value string
+	}{
+		{input: "a=b", key: "a", value: "b"},
+		{input: "a = b", key: "a", value: "b"},
+		{input: "a = b = c", key: "a", value: "b = c"},
+	}
+
+	for _, p := range params {
+		actualKey, actualValue := SplitFormat(p.input)
+		if actualKey != p.key || actualValue != p.value {
+			t.Errorf("SplitFormat() = %s => %s, want %s => %s", actualKey, actualValue, p.key, p.value)
+		}
+	}
+}
